Add files via upload
My goal is to wrangle WeRateDogs Twitter data to create interesting and trustworthy analyses and visualizations. I had three datasets which I acquired by different means. The first dataset 'twitter_Archive.csv' was a CSV file which was provided by the organizers and I downloaded it into my local machine, the second dataset which is 'image-predictions.tsv', a TSV file that I also downloaded using a url link provided by the organizers and the third dataset 'tweet_json' was a json file also provided by the organizer. I had to make use of this json file because I was granted access to scrap data on twitter.